requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
psycopg2-binary==2.9.11
PyJWT==2.8.0
rq==1.10.1
redis==7.1.0
alembic==1.11.1
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.11
